Fixes underflow error in squared_euclid_distance

+++ Maze/Common/grid.rs
@@ -9,7 +9,7 @@ pub type Position = (usize, usize);
 
 /// Returns the squared euclidian distance between two `Position`s
 pub fn squared_euclidian_distance(p1: &Position, p2: &Position) -> usize {
-    (p1.0 - p2.0).pow(2) + (p1.1 - p2.1).pow(2)
+    ((p1.0 as isize - p2.0 as isize).pow(2) + (p1.1 as isize - p2.1 as isize).pow(2)) as usize
 }
 
 /// Represents a wrapper type for a 2D array with added functionality
